requirements.txt
streamlit
pandas
psycopg2-binary
geopy
folium
streamlit-folium
httpx
requests
# ssl 
# date 
dotenv
# os